requirements.txt
Flask==2.0.2
pymongo==3.12.1
requests==2.26.0
opentelemetry-api==1.19.0
opentelemetry-distro==0.40b0
opentelemetry-instrumentation==0.40b0
opentelemetry-exporter-otlp==1.19.0
